catword/requirements.txt
android==0.0.1
gevent_socketio==0.3.6
numpy==1.16.2
remi==2021.3.2
tensorflow==2.6.0
tensorflow_cpu==2.5.0
